Refactor commands and model into separate sub-packages
