+++ clowder/commands/fix.py
