+++ clowder/commands/litter.py
