+++ clowder/commands/meow.py
