+++ clowder/commands/play.py
