+++ clowder/commands/purr.py
