+++ clowder/model/group.py
@@ -1,0 +1,5 @@
+class Group(object):
+
+    def __init__(self, name, projects):
+        self.name = name
+        self.projects = projects
