+++ clowder/model/manifest.py
