+++ clowder/model/parser.py
@@ -1,0 +1,15 @@
+class Parser(object):
+
+    def __init__(self, path):
+        self.path = path
+        # self.defaultRef = ref
+        # self.defaultRemote = remote
+        # self.remotes = remotes
+        # self.allGroups = allGroups
+        # self.currentGroups = currentGroups
+
+    def getGroups(self):
+        pass
+
+    def getSnapshots(self):
+        pass
