+++ clowder/model/project.py
@@ -1,0 +1,7 @@
+class Project(object):
+
+    def __init__(self, name, projectPath):
+        self.name = name
+        self.path = projectPath
+        self.ref = ref
+        self.remote = remote
