+++ clowder/model/remote.py
@@ -1,0 +1,6 @@
+class Remote(object):
+
+    def __init__(self, name, url):
+        self.name = name
+        self.sshUrl = url # TODO: Properly implement
+        self.httpsUrl = url # TODO: Properly implement
